% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_utilities.R
\name{ameans}
\alias{ameans}
\title{Fast Marginal Mean over an Array Dimension}
\usage{
ameans(a, i)
}
\arguments{
\item{a}{A numeric array of arbitrary dimension.}

\item{i}{Integer. The dimension over which to compute the marginal means.}
}
\value{
A numeric array with the \code{i}-th dimension collapsed via averaging.
}
\description{
Computes the marginal means over a specified dimension of a multi-dimensional array.
This is a faster alternative to \code{apply(an_array, i, mean)} when \code{i} is a single dimension.
}
\details{
This function reorders the array dimensions using \code{aperm()} to bring the target dimension to the end,
and then efficiently applies \code{rowMeans()} to compute marginal means. It is especially efficient for large arrays
compared to the base \code{apply()} function.
}
\examples{
x <- array(rnorm(2 * 3 * 4), dim = c(2, 3, 4))
apply(x, c(1, 3), mean)  # base R version
ameans(x, 2)             # faster version over the second dimension

}
\seealso{
\code{\link[base]{apply}}, \code{\link[base]{aperm}}, \code{\link[base]{rowMeans}}
}
