% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_utilities.R
\name{vec}
\alias{vec}
\title{Vectorize a Matrix}
\usage{
vec(x, mode = "any")
}
\arguments{
\item{M}{A numeric matrix.}
}
\value{
A numeric vector.
}
\description{
Converts a matrix to a column-major vector.
}
